@@ -1,13 +1,11 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 33
 
     defaultConfig {
-        minSdkVersion 17
-        targetSdkVersion 29
-        versionCode 13
-        versionName "1.3"
+        minSdkVersion 19
+        targetSdkVersion 33
     }
     buildTypes {
         release {
@@ -15,6 +13,7 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    namespace 'com.nbsp.materialfilepicker'
 }
 
 dependencies {
