@@ -26,7 +26,7 @@ import java.util.ArrayList;
 public class ProductDetailsFragment extends Fragment {
 
     WebView webView;
-    TextView textJobQualification, textWorkDay, textWorkTime, textExp, textType,marital, text_desc,textSkills;
+    TextView text_negotiate, text_desc, text_website;
     ItemProduct itemJob;
 
     Button btn_call, btn_whatsapp,btn_back;
@@ -43,81 +43,51 @@ public class ProductDetailsFragment extends Fragment {
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_job_details, container, false);
+        View rootView = inflater.inflate(R.layout.fragment_product_details, container, false);
         if (getArguments() != null) {
             itemJob = (ItemProduct) getArguments().getSerializable("itemJob");
         }
-        text_desc= rootView.findViewById(R.id.text_desc);
-        webView = rootView.findViewById(R.id.text_job_description);
-        textJobQualification = rootView.findViewById(R.id.text_job_qualification);
-        textWorkDay = rootView.findViewById(R.id.text_job_work_day);
-        textWorkTime = rootView.findViewById(R.id.text_job_work_time);
-        textExp = rootView.findViewById(R.id.text_job_exp);
-        textType = rootView.findViewById(R.id.text_job_type);
-        textSkills= rootView.findViewById(R.id.textSkills);
-        marital = rootView.findViewById(R.id.marital);
+        mSkills = new ArrayList<>();
+
+        text_negotiate = rootView.findViewById(R.id.text_negotiate);
+        text_desc = rootView.findViewById(R.id.text_desc);
+        text_website = rootView.findViewById(R.id.text_website);
+
         btn_call = rootView.findViewById(R.id.btn_call);
         btn_whatsapp = rootView.findViewById(R.id.btn_whats);
         btn_back = rootView.findViewById(R.id.btn_back);
-        mSkills = new ArrayList<>();
 
-//        recyclerView.setHasFixedSize(true);
-//        recyclerView.setLayoutManager(new LinearLayoutManager(getActivity(), LinearLayoutManager.VERTICAL, false));
-//        recyclerView.setFocusable(false);
-//        recyclerView.setNestedScrollingEnabled(false);
-        text_desc.setText(itemJob.getProductDesc());
-        textWorkDay.setText(itemJob.getAge());
-        textWorkTime.setText(itemJob.getSex());
-        textType.setText(itemJob.getProductType());
-        marital.setText(itemJob.getMarital());
 
-    btn_back.setOnClickListener(new View.OnClickListener() {
-        @Override
-        public void onClick(View view) {
-getActivity().onBackPressed();
-    }
-    });
-        btn_call.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(Intent.ACTION_DIAL, Uri.fromParts("tel", itemJob.getProductPhoneNumber(), null));
-                startActivity(intent);
+        btn_back.setOnClickListener(view -> getActivity().onBackPressed());
+        btn_call.setOnClickListener(view -> {
+            Intent intent = new Intent(Intent.ACTION_DIAL, Uri.fromParts("tel", itemJob.getProductPhoneNumber(), null));
+            startActivity(intent);
+        });
+        btn_whatsapp.setOnClickListener(view -> {
+            PackageManager pm= getActivity().getPackageManager();
+            try {
+                Intent waIntent = new Intent(Intent.ACTION_SEND);
+                waIntent.setType("text/plain");
+                String text =
+                        itemJob.getProductName() + "\n" +
+                                getString(R.string.job_company_lbl) + itemJob.getProductCompanyName() + "\n" +
+                                getString(R.string.job_designation_lbl) + itemJob.getProductDesignation() + "\n" +
+                                getString(R.string.job_phone_lbl) + itemJob.getProductPhoneNumber() + "\n" +
+                                getString(R.string.job_address_lbl) + itemJob.getCity() + "\n\n" +
+                                "Download Application here https://play.google.com/store/apps/details?id=com.jellysoft.sundigitalindia";
+                PackageInfo info=pm.getPackageInfo("com.whatsapp", PackageManager.GET_META_DATA);
+                //Check if package exists or not. If not then code
+                //in catch block will be called
+                waIntent.setPackage("com.whatsapp");
+
+                waIntent.putExtra(Intent.EXTRA_TEXT, text);
+                startActivity(Intent.createChooser(waIntent, "Share with"));
+
+            } catch (PackageManager.NameNotFoundException e) {
+                Toast.makeText(getActivity(), "WhatsApp not Installed", Toast.LENGTH_SHORT)
+                        .show();
             }
         });
-        btn_whatsapp.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                PackageManager pm= getActivity().getPackageManager();
-                try {
-
-                    Intent waIntent = new Intent(Intent.ACTION_SEND);
-                    waIntent.setType("text/plain");
-                    String text =
-                            itemJob.getProductName() + "\n" +
-                                    getString(R.string.job_company_lbl) + itemJob.getProductCompanyName() + "\n" +
-                                    getString(R.string.job_designation_lbl) + itemJob.getProductDesignation() + "\n" +
-                                    getString(R.string.job_phone_lbl) + itemJob.getProductPhoneNumber() + "\n" +
-                                    getString(R.string.job_address_lbl) + itemJob.getCity() + "\n\n" +
-                                    "Download Application here https://play.google.com/store/apps/details?id=com.jellysoft.sundigitalindia";
-                    PackageInfo info=pm.getPackageInfo("com.whatsapp", PackageManager.GET_META_DATA);
-                    //Check if package exists or not. If not then code
-                    //in catch block will be called
-                    waIntent.setPackage("com.whatsapp");
-
-                    waIntent.putExtra(Intent.EXTRA_TEXT, text);
-                    startActivity(Intent.createChooser(waIntent, "Share with"));
-
-                } catch (PackageManager.NameNotFoundException e) {
-                    Toast.makeText(getActivity(), "WhatsApp not Installed", Toast.LENGTH_SHORT)
-                            .show();
-                }
-            }
-        });
-//        if (!itemJob.getJobSkill().isEmpty()) {
-//            mSkills = new ArrayList<>(Arrays.asList(itemJob.getJobSkill().split(",")));
-//            SkillsAdapter skillsAdapter = new SkillsAdapter(getActivity(), mSkills);
-//            recyclerView.setAdapter(skillsAdapter);
-//        }
 
         return rootView;
     }
