@@ -15,7 +15,9 @@ import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.adapter.JobAdapter;
+import com.example.adapter.ProductAdapter;
 import com.example.item.ItemJob;
+import com.example.item.ItemProduct;
 import com.example.util.API;
 import com.example.util.Constant;
 import com.example.util.EndlessRecyclerViewScrollListener;
@@ -42,8 +44,8 @@ import cz.msebera.android.httpclient.Header;
 
 public class LatestProduct extends AppCompatActivity {
     public RecyclerView recyclerView;
-    ArrayList<ItemJob> mListItem;
-    JobAdapter adapter;
+    ArrayList<ItemProduct> mListItem;
+    ProductAdapter adapter;
     private ProgressBar progressBar;
     private LinearLayout lyt_not_found;
     boolean isFirst = true, isOver = false;
@@ -108,7 +110,6 @@ public class LatestProduct extends AppCompatActivity {
         });
     }
     private void getLatestOrRecent() {
-
         AsyncHttpClient client = new AsyncHttpClient();
         RequestParams params = new RequestParams();
         JsonObject jsObj = (JsonObject) new Gson().toJsonTree(new API());
@@ -133,24 +134,27 @@ public class LatestProduct extends AppCompatActivity {
                 try {
                     JSONObject mainJson = new JSONObject(result);
                     JSONArray jsonArray = mainJson.getJSONArray(Constant.ARRAY_NAME);
-                    JSONObject objJson;
+                    JSONObject jsonObject;
                     if (jsonArray.length() > 0) {
                         for (int i = 0; i < jsonArray.length(); i++) {
-                            objJson = jsonArray.getJSONObject(i);
-                            if (objJson.has(Constant.STATUS)) {
+                            jsonObject = jsonArray.getJSONObject(i);
+                            if (jsonObject.has(Constant.STATUS)) {
                                 lyt_not_found.setVisibility(View.VISIBLE);
                             } else {
-                                ItemJob objItem = new ItemJob();
-                                objItem.setJobLogo(objJson.getString(Constant.PRODUCT_LOGO));
-                                objItem.setId(objJson.getString(Constant.PRODUCT_ID));
-                                objItem.setJobType(objJson.getString(Constant.PRODUCT_TYPE));
-                                objItem.setJobName(objJson.getString(Constant.PRODUCT_NAME));
-                                objItem.setJobCategoryName(objJson.getString(Constant.CATEGORY_NAME));
-                                objItem.setCity(objJson.getString(Constant.CITY_NAME));
-                                objItem.setJobDate(objJson.getString(Constant.PRODUCT_START_DATE));
-                                objItem.setpLate(objJson.getString(Constant.PRODUCT_END_DATE));
-                                objItem.setViews(objJson.getString("views"));
-                                objItem.setJobFavourite(objJson.getBoolean(Constant.PRODUCT_FAVOURITE));
+                                ItemProduct objItem = new ItemProduct();
+                                objItem.setProductLogo(jsonObject.getString(Constant.PRODUCT_LOGO));
+                                objItem.setId(jsonObject.getString(Constant.PRODUCT_ID));
+                                objItem.setProductType(jsonObject.getString(Constant.PRODUCT_TYPE));
+                                objItem.setProductName(jsonObject.getString(Constant.PRODUCT_NAME));
+                                objItem.setProductPhoneNumber(jsonObject.getString(Constant.PRODUCT_PHONE_NO));
+                                objItem.setProductPhoneNumber2(jsonObject.getString(Constant.PRODUCT_PHONE_NO2));
+                                objItem.setProductCategoryName(jsonObject.getString(Constant.CATEGORY_NAME));
+                                objItem.setProductPrice(jsonObject.getString(Constant.PRODUCT_PRICE));
+                                objItem.setProductSellingPrice(jsonObject.getString(Constant.PRODUCT_SELLING_PRICE));
+                                objItem.setCity(jsonObject.getString(Constant.CITY_NAME));
+                                objItem.setProductDate(jsonObject.getString(Constant.PRODUCT_START_DATE));
+                                objItem.setPLate(jsonObject.getString(Constant.PRODUCT_END_DATE));
+                                objItem.setViews(jsonObject.getString("views"));
                                 mListItem.add(objItem);
                             }
                         }
@@ -172,7 +176,6 @@ public class LatestProduct extends AppCompatActivity {
                 showProgress(false);
                 lyt_not_found.setVisibility(View.VISIBLE);
             }
-
         });
     }
 
@@ -180,11 +183,10 @@ public class LatestProduct extends AppCompatActivity {
         if (mListItem.size() == 0) {
             lyt_not_found.setVisibility(View.VISIBLE);
         } else {
-
             lyt_not_found.setVisibility(View.GONE);
             if (isFirst) {
                 isFirst = false;
-                adapter = new JobAdapter(LatestProduct.this, mListItem);
+                adapter = new ProductAdapter(LatestProduct.this, mListItem);
                 recyclerView.setAdapter(adapter);
             } else {
                 adapter.notifyDataSetChanged();
@@ -194,7 +196,7 @@ public class LatestProduct extends AppCompatActivity {
                 @Override
                 public void onItemClick(int position) {
                     String jobId = mListItem.get(position).getId();
-                    Intent intent = new Intent(LatestProduct.this, JobDetailsActivity.class);
+                    Intent intent = new Intent(LatestProduct.this, ProductDetailsActivity.class);
                     intent.putExtra("Id", jobId);
                     startActivity(intent);
                 }
@@ -228,7 +230,6 @@ public class LatestProduct extends AppCompatActivity {
     public void getSaveJob(Events.SaveJob saveJob) {
         for (int i = 0; i < mListItem.size(); i++) {
             if (mListItem.get(i).getId().equals(saveJob.getJobId())) {
-                mListItem.get(i).setJobFavourite(saveJob.isSave());
                 adapter.notifyItemChanged(i);
             }
         }
